Bug: Edimax BLE Dongle Fails After Teardown and Re-Instantiation
This patch addresses an issue where the some RTK BLE dongles fail to perform
an HCI reset after the transport is torn down and re-instantiated. To address
that, we prevent crashing the background threads when invalid data comes in,
and handle resetting the parser. A retry for the HCI reset has been implemented
for the RTK dongles where errant data comes in after being torn down.